requirements.txt
django==3.2
djangorestframework
django-filter
drf-yasg

